parallelize utils

+++ src/utils/MRI_PostProcessing.py
@@ -1,16 +1,21 @@
+import os
 import ants
 import pickle
+import random
 import numpy as np
 from tqdm import tqdm
-import nibabel as nib
 from ..config import MRI2PETConfig
 
 config = MRI2PETConfig()
 mri_list = pickle.load(open('./src/data/mriDataset.pkl', 'rb'))
+random.shuffle(mri_list)
 
 for mri_fpath in tqdm(mri_list):
-    img = ants.image_read(fpath)
+    fpath = mri_fpath.replace('.nii.gz', '.npy')
+    if os.path.exists(fpath):
+        continue
+
+    img = ants.image_read(mri_fpath)
     data = img.numpy()
     data = (data - data.min()) / (data.max() - data.min())
-    fpath = mri_fpath.replace('.nii.gz', '.npy')
     np.save(fpath, data)
